+++ changelog.md
@@ -7,3 +7,4 @@
 * [added] preview
 * [added] cell tracking
 * [added] export to spreadsheet file
+* [added] Debug images